+++ index.js
@@ -1,9 +1,0 @@
-'use strict';
-
-const path = require('path');
-
-module.exports = {
-  name: 'graphitix',
-  entryHtml: path.join(__dirname, 'index.html'),
-  rootDir: __dirname,
-};