+++ src/adder.js
@@ -1,7 +1,0 @@
-'use strict';
-
-function add(a, b){
-  return Number(a) + Number(b);
-}
-
-module.exports = { add };